java

+++ RegularCleaning.java
@@ -5,7 +5,7 @@ import java.io.*;
 class RegularCleaning {
     public static void main(String[] args) throws java.lang.Exception {
         Scanner sc = new Scanner(System.in);
-        int N = sc.nextInt();
+        ll N = sc.nextInt();
         int c = 0;
 
         c = 10 - (N % 10);
